Implementing persistent_map - persistent binary tree
This is a persistent binary tree data structure that behaves std::map<key,value>-ish
This will be helpful for implementing delta blocks. In particular, this will allow to:
- share transaction data between delta blocks
- keep a spent-index for each delta block to cross check them for mergeability

Comes with tests.

The first change needs an abstraction of the container-structure of CBlock to hide it
behind an appropriate API and this will be one of the following commits.

Substitute std::vector in CBlock with a persistent_map based implementation

Rename NODE_WEAKBLOCKS to NODE_DELTABLOCKS

CDeltaBlock - delta blocks implementation

These are the beginnings of and the first incomplete demo of
the deltablocks idea.

Switch to an aggressive 72s mempool expiry

This is as to not have cruft pile up in the mempool during testing.