Block as abstract transaction containers
This change abstracts the implementation details of a CBlock's
transaction set, which is currently simply a
std::vector<CTransactionRef> away from a CBlock's interface.
This will allow the internal representation of a CBlock to bee
more easily changed for different storage models and container
types (such as a Merklix-aware binary tree, or a mmap()ed on disk
representation, potentially aiding an implementation of BUIP120
as well). It should also help to prepare for further potential
changes to BCH down the road (such as Merklix trees) as well as
making dealing with transaction order (CTOR) easier and more
localized.