Remove dsproof's from beta
* remove the beta tag for the dsproof message

Also, cleanup protocol.h and .cpp. Assign hard coded numbers to the enum
for message types. This makes it easier to read and also prevents any future
conflicts with other projects since there is no need to keep messages in some
sort of special order in the enum field. In this way new messages can be
easily added or removed over time.